+++ Ladybug/Model/Radar.swift
@@ -11,9 +11,18 @@ import Foundation
 class Radar: Codable {
     let id: RadarID
     var metadata: RadarMetadata?
+    let firstViewedDate: Date
+    var lastViewedDate: Date
+    var favoritedDate: Date? = nil
 
     init(id: RadarID, metadata: RadarMetadata? = nil) {
         self.id = id
         self.metadata = metadata
+        firstViewedDate = Date()
+        lastViewedDate = Date()
+    }
+
+    var idString: String {
+        return String(id.id)
     }
 }